fix:when u shut down the client while never click login to load the config, the null config will be saved

+++ src/main/java/com/prohect/sql_frontend/main/login/ClientConfig.java
@@ -56,6 +56,7 @@ public class ClientConfig {
     }
 
     public static void saveConfig(ClientConfig clientConfig) {
+        if (clientConfig == null) return;
         try {
             configFile.delete();
             configFile.createNewFile();
